finito?

+++ server_client/msgparsing.c
@@ -32,7 +32,7 @@ void parse(char *src, char* dest, int offset){
 
     int delim_index = has_delim(src, offset);
     if(delim_index != -1){
-        if (delim_index != strlen(aux) - 3){
+        if ((size_t)delim_index != strlen(aux) - 3){
             printf("Errore, delimitatore presente nel messaggio\n"); //lasciare o no?
             return;
         }
